requirements.txt
# Core Dependencies
numpy>=1.21.0
scipy>=1.7.0

# Audio Processing & Diarization
pyannote.audio>=3.0.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0

# Video Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0  # For facial landmarks and lip movement detection

# Transcript Processing
pysrt>=1.1.2
webvtt-py>=0.4.6

# ML/NLP for Name Extraction
transformers>=4.30.0
spacy>=3.5.0

# Output Generation
jinja2>=3.1.0

# CLI and Configuration
click>=8.1.0
pyyaml>=6.0
tqdm>=4.65.0  # Progress bars

# Utilities
python-dateutil>=2.8.0
pillow>=10.0.0

# Optional: LLM Integration (can be disabled)
openai>=1.0.0  # Optional, for name extraction assistance
anthropic>=0.18.0  # Optional, for name extraction assistance

# Development Dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
